+++ Horizon/Horizon/Sources/Features/Assist/AssistChat/Domain/AssistChatInteractor.swift
@@ -528,3 +528,18 @@ struct AssistChatInteractorPreview: AssistChatInteractor {
     func publish(action: AssistChatAction) {}
     var listen: AnyPublisher<AssistChatResponse, Error> = Empty().eraseToAnyPublisher()
 }
+
+extension API {
+    func makeRequest<Request: APIRequestable>(_ requestable: Request) -> AnyPublisher<Request.Response?, Error> {
+        AnyPublisher<Request.Response?, Error> { [weak self] subscriber in
+            self?.makeRequest(requestable) { response, _, error in
+                if let error = error {
+                    subscriber.send(completion: .failure(error))
+                    return
+                }
+                subscriber.send(response)
+            }
+            return AnyCancellable { }
+        }
+    }
+}
